CensRCalc: Analytical Censoring Parameter Determination

Calculations for the expected censoring proportion under independent censoring to determine a censoring distribution parameter that achieves a target censoring proportion in simulation studies. The expected censoring proportion is evaluated by numerical integration over time and covariates, with support for administrative censoring and accrual periods.

Details

Let the event time be \(T^*\mid\mathbf{X}\sim F_{T^*\mid\mathbf{X}}\) with survival function \(S_{T^*\mid\mathbf{X}}=1-F_{T^*\mid\mathbf{X}}\) and the random censoring time be \(C_{\text{rnd}}\mid\lambda_{C_{\text{rnd}}}\sim F_{C_{\text{rnd}} \mid\lambda_{C_{\text{rnd}}}}\) with survival function \(S_{C_{\text{rnd}}\mid\lambda_{C_{\text{rnd}}}}=1-F_{C_{\text{rnd}} \mid\lambda_{C_{\text{rnd}}}}\), where \(\lambda_{C_{\text{rnd}}}\) is the random censoring parameter of interest. Administrative censoring occurs at time \(\tau_{\text{adm}}\). Accrual takes place over \([0,\tau_{\text{acc}}]\) where the random accrual time is \(R\sim U(0,\tau_{\text{acc}})\). The equivalent random administrative censoring time after accrual is \(C_{\text{adm}}=\tau_{\text{adm}}-R\). Hence, \(C_{\text{adm}}\sim U(\tau_{\text{adm}}-\tau_{\text{acc}}, \tau_{\text{adm}})\).

Under these assumptions, the probability of censoring dependent on \(\lambda_{C_{\text{rnd}}}\) and the covariate vector \(\mathbf{X}\) is

\[ \begin{aligned} P(\delta=0\mid \mathbf{X},\lambda_{C_{\text{rnd}}}) = &\int_{0}^{\tau_{\text{adm}} - \tau_{\text{acc}}} S_{T^{*}\mid\mathbf{X}}(c_{\text{rnd}}) f_{C_{\text{rnd}}\mid\lambda_{C_{\text{rnd}}}}(c_{\text{rnd}})dc_{\text{rnd}} + \\ &\int_{\tau_{\text{adm}} - \tau_{\text{acc}}}^{\tau_{\text{adm}}} S_{T^{*}\mid\mathbf{X}}(c_{\text{rnd}})S_{C_{\text{adm}}}(c_{\text{rnd}}) f_{C_{\text{rnd}}\mid\lambda_{C_{\text{rnd}}}}(c_{\text{rnd}})dc_{\text{rnd}} + \\ &\int_{\tau_{\text{adm}} - \tau_{\text{acc}}}^{\tau_{\text{adm}}} S_{T^{*}\mid\mathbf{X}}(C_{\text{adm}}) S_{C_{\text{rnd}}\mid\lambda_{C_{\text{rnd}}}}(C_{\text{adm}}) f_{C_{\text{adm}}}(C_{\text{adm}})dc_{\text{adm}} \end{aligned} \]

Next we take the expectation over the covariate distribution to get the expected censoring proportion: \[P(\delta=0\mid\lambda_{C_{\text{rnd}}})=E_X\left[P(\delta=0\mid \mathbf{X},\lambda_{C_{\text{rnd}}})\right].\] This is implemented in estimate_cens_prop(). To obtain the censoring parameter \(\lambda_{C_{\text{rnd}}}\) that achieves a target censoring proportion \(p_C\), we solve the root finding problem \[P(\delta=0\mid\lambda_{C_{\text{rnd}}}) - p_C = 0\]

Special cases include no accrual time (\(\tau_{\text{acc}}=0\)) and no administrative censoring (\(\tau_{\text{adm}}=\infty\)). See the package vignette vignette("full_details", package = "CensRCalc") for details.

Numerically, the integrals are evaluated using adaptive cubature (calculus::integral()), which was extended to handle discrete and mixed discrete/continuous covariates.

See also

Author

Maintainer: Lukas Klein lukas.klein@h-da.de

Authors:

Lukas Klein lukas.klein@h-da.de

Henrik Stahl henrik.stahl@h-da.de

Other contributors:

Antje Jahn-Eimermacher [contributor]

Gunter Grieser [contributor]